ant virus/requirements.txt
requests==2.31.0
python-magic==0.4.27
cryptography==41.0.7
